refactor: 배포한 모달 패키지 사용
